# packages/contracts/openapi/openapi.json
{
  "components": {
    "parameters": {
...
      "Origin": {
        "in": "header",
        "name": "Origin",
        "required": false,
        "schema": {
          "format": "uri",
          "type": "string"
        }
      },
...
      "Referer": {
        "in": "header",
        "name": "Referer",
        "required": false,
        "schema": {
          "format": "uri",
          "type": "string"
        }
      }
    },
    "responses": {
...
      "StandardError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Standard error envelope"
      }
    },
    "schemas": {
...
      "AuthLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "AuthTokenResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AuthTokenData"
          },
          "message": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "trace_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "refreshCookie": {
        "description": "Secure, HttpOnly, SameSite refresh cookie; default lifetime seven days.",
        "in": "cookie",
        "name": "refresh_token",
        "type": "apiKey"
      },
      "serviceBearerAuth": {
        "bearerFormat": "JWT",
        "description": "Internal HS256 service JWT; fixed issuer/audience/scope and short lifetime.",
        "scheme": "bearer",
        "type": "http",
        "x-ai-caller-subject": "ai-api",
        "x-required-audience": "business-api",
        "x-required-issuers": [
          "ai-api",
          "monitoring"
        ],
        "x-required-scope": "health"
      }
    }
  },
  "info": {
    "description": "Sprint 2 Requirement/Baseline R4 implementation candidate authorized by PORTFOLIO-P1-RUNTIME-01. It remains pending Review and the shared real-runtime Gate; persistence adapters and FLOW remain disabled.",
    "summary": "Sprint 2 Requirement/Baseline R4 implementation candidate",
    "title": "AI Product Design and Validation API",
    "version": "0.1.0-dev",
    "x-contract-reviewers": [
      "frontend",
      "ai-data",
      "review"
    ],
    "x-contract-status": "candidate"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful response",
            "headers": {
              "Set-Cookie": {
                "description": "Sets the rotated Secure, HttpOnly, SameSite refresh cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [],
        "tags": [
          "identity"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-implementation-status": "implemented-candidate",
        "x-refresh-cookie": {
          "action": "set",
          "httpOnly": true,
          "maxAgeSeconds": 604800,
          "sameSite": "Strict-or-Lax-by-deployment",
          "secure-in-production": true
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Referer"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful response",
            "headers": {
              "Set-Cookie": {
                "description": "Clears the refresh cookie with an expired Max-Age.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "refreshCookie": []
          }
        ],
        "tags": [
          "identity"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-cookie-command-origin-policy": {
          "csrfToken": "required-when-deployment-enables-double-submit",
          "failureCodes": [
            "ORIGIN_REQUIRED",
            "ORIGIN_MISMATCH",
            "CSRF_INVALID"
          ],
          "requireOneOf": [
            "Origin",
            "Referer"
          ],
          "sameOrigin": true
        },
        "x-implementation-status": "implemented-candidate",
        "x-refresh-cookie": {
          "action": "clear",
          "httpOnly": true,
          "maxAgeSeconds": 0,
          "sameSite": "Strict-or-Lax-by-deployment",
          "secure-in-production": true
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotates the seven-day refresh token. Reuse revokes the complete token family.",
        "operationId": "refreshAccessToken",
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          },
          {
            "$ref": "#/components/parameters/Referer"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            },
            "description": "Successful response",
            "headers": {
              "Set-Cookie": {
                "description": "Sets the rotated Secure, HttpOnly, SameSite refresh cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "tags": [
          "identity"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-cookie-command-origin-policy": {
          "csrfToken": "required-when-deployment-enables-double-submit",
          "failureCodes": [
            "ORIGIN_REQUIRED",
            "ORIGIN_MISMATCH",
            "CSRF_INVALID"
          ],
          "requireOneOf": [
            "Origin",
            "Referer"
          ],
          "sameOrigin": true
        },
        "x-implementation-status": "implemented-candidate",
        "x-refresh-cookie": {
          "action": "rotate",
          "httpOnly": true,
          "maxAgeSeconds": 604800,
          "sameSite": "Strict-or-Lax-by-deployment",
          "secure-in-production": true
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            },
            "description": "Successful response",
            "headers": {
              "Set-Cookie": {
                "description": "Sets the rotated Secure, HttpOnly, SameSite refresh cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [],
        "tags": [
          "identity"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-implementation-status": "implemented-candidate",
        "x-refresh-cookie": {
          "action": "set",
          "httpOnly": true,
          "maxAgeSeconds": 604800,
          "sameSite": "Strict-or-Lax-by-deployment",
          "secure-in-production": true
        }
      }
    },
    "/api/v1/file-versions/{version_id}/relations": {
      "post": {
        "operationId": "createFileRelation",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileRelationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileRelationResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "files"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-candidate"
      }
    },
    "/api/v1/file-versions/{version_id}:download": {
      "post": {
        "operationId": "createFileVersionDownload",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadFileVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadFileVersionResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "files"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-implementation-status": "implemented-candidate"
      }
    },
    "/api/v1/files/uploads": {
      "post": {
        "operationId": "initFileUpload",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileUploadInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "files"
        ],
        "x-canonical-event-transaction": {
          "event_name": "file.upload.started",
          "only_when_frozen_canonical_event_exists": true,
          "required": true
        },
        "x-contract-phase": "sprint-1-candidate",
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-candidate",
        "x-pending-file-version": {
          "forbiddenBeforeComplete": [
            "list",
            "download",
            "relate",
            "set-current"
          ],
          "visibility": "internal-only"
        }
      }
    },
    "/api/v1/files/uploads/{upload_id}:abort": {
      "post": {
        "operationId": "abortFileUpload",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortFileUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbortFileUploadResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "files"
        ],
        "x-contract-phase": "sprint-1-candidate",
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-candidate",
        "x-pending-file-version": {
          "auditable": true,
          "businessSuccess": false,
          "orphanCleanupCandidate": true,
          "terminalStates": [
            "aborted",
            "failed"
          ]
        }
      }
    },
    "/api/v1/files/uploads/{upload_id}:complete": {
      "post": {
        "description": "One-time pending-to-available finalization. Validates actor, expiry, object state, size, MIME and checksum. Only the successful metadata transaction may set current_version_id, create a relation, audit, Outbox and idempotency result.",
        "operationId": "completeFileUpload",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteFileUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/StandardError"
          },
          "401": {
            "$ref": "#/components/responses/StandardError"
          },
          "403": {
            "$ref": "#/components/responses/StandardError"
          },
          "404": {
            "$ref": "#/components/responses/StandardError"
          },
          "409": {
            "$ref": "#/components/responses/StandardError"
          },
          "422": {
            "$ref": "#/components/responses/StandardError"
          },
          "500": {
            "$ref": "#/components/responses/StandardError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "files"
        ],
        "x-canonical-event-transaction": {
          "event_name": "file.upload.completed",
          "only_when_frozen_canonical_event_exists": true,
          "required": true
        },
        "x-contract-phase": "sprint-1-candidate",
        "x-file-version-finalization": {
          "finalObjectStrategy": "server-conditional-copy",
          "from": "pending",
          "immutableAfter": true,
          "once": true,
          "persistedImmutableIdentifier": "storage_version_id",
          "to": "available",
          "uploadChecksumHeader": "x-amz-checksum-sha256"
        },
        "x-idempotency-scope": "actor+endpoint+request-hash",
        "x-implementation-status": "implemented-candidate"
      }
    },
...
  },
...
}